Add configure option --disable-card-manager and remove switch to gpg2.
* configure.ac: Add option --disable-card-manager.
* src/cardman.h [!ENABLE_CARD_MANAGER]: Do not define anything.
* src/clipboard.c (file_open): Cast off_t.
* src/gpa.c (option_entries) [!ENABLE_CARD_MANAGER]: Remove option
"--card".
(quit_if_no_window) [!ENABLE_CARD_MANAGER]: Do not check for open card
manager.
(gpa_open_cardmanager) [!ENABLE_CARD_MANAGER]: Disable.
(main): Do not call gpa_switch_to_gpg2; this is now done by gpgme.  Do
not call gpa_open_cardmanager.
* src/gpa.h (gpa_windows_menu_action_entries) [!ENABLE_CARD_MANAGER]:
Remove card manger entry.
* src/gpgmetools.c (gpa_switch_to_gpg2): Remove.
* src/keytable.c (first_half_done_cb): Change wording of notice for
missing gpgsm.
* src/server.c (cmd_start_cardmanager) [!ENABLE_CARD_MANAGER]:
Disable.
(register_commands) [!ENABLE_CARD_MANAGER]: Do not register.

+++ src/clipboard.c
@@ -427,9 +427,10 @@ file_open (GtkAction *action, gpointer param)
 
       /* TRANSLATORS: The arguments are the filename, the integer size
 	 and the unit (such as KB or MB).  */
-      str = g_strdup_printf (_("The file %s is %lli%s large.  Do you really "
+      str = g_strdup_printf (_("The file %s is %llu%s large.  Do you really "
 			       " want to open it?"), filename,
-			       buf.st_size / 1024 / 1024, "MB");
+                             (unsigned long long)buf.st_size / 1024 / 1024,
+                             "MB");
       labelMessage = gtk_label_new (str);
       g_free (str);
       gtk_label_set_line_wrap (GTK_LABEL (labelMessage), TRUE);
